+++ src/Telemetry/Library/InternalsVisibleToTests.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Microsoft.Quantum.Telemetry.Tests")]
+
+namespace Microsoft.Quantum.Telemetry
+{
+}
